+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.17.6] - 2018-03-01
+### Changed
+- config.js, better balance of units spawning
+- prototype.spawn.js, refactored Spawning routine, unified into spawnCustomCreep() method
+                      hardcoded nearby rooms, temporary fix
+- role.hero.js, bug fix
+
 ## [2.17.5] - 2018-03-01
 ### Changed
 - prototype.creep.js, renamed recharge() method into getEnergy()